Move Vue setup to utility function
